+++ lib/src/api/connection_manager.dart
@@ -3,7 +3,7 @@ import 'package:ledger_flutter_plus/ledger_flutter_plus.dart';
 abstract class ConnectionManager {
   Future<void> connect(LedgerDevice device, {LedgerOptions? options});
 
-  Future<void> disconnect(LedgerDevice device);
+  Future<void> disconnect(String deviceId);
 
   Future<T> sendOperation<T>(
     LedgerDevice device,